79: PR review updates
- Update yup schema
- Move mealcards and related code to common
- Add created contion before redirection from plan page
- Make all routes suspense
