+++ src/App.tsx
@@ -26,32 +26,27 @@ function App() {
         <Provider store={store}>
           <BrowserRouter>
             <NavBar />
-            <Routes>
-              <Route path="/" element={<div>Hello world!</div>} />
+            <Suspense>
+              <Routes>
+                <Route path="/" element={<div>Hello world!</div>} />
 
-              <Route element={<ProtectedRoute />}>
-                <Route path="planner" element={<Planner />}>
-                  <Route index element={<div>Planner</div>} />
-                  <Route path="ingredients" element={<Ingredients />} />
-                  <Route path="dishes" element={<Dishes />} />
-                  <Route path="plans">
-                    <Route index element={<Plans />} />
-                    <Route
-                      path="edit/:id"
-                      element={
-                        <Suspense>
-                          <EditPlan />
-                        </Suspense>
-                      }
-                    />
+                <Route element={<ProtectedRoute />}>
+                  <Route path="planner" element={<Planner />}>
+                    <Route index element={<div>Planner</div>} />
+                    <Route path="ingredients" element={<Ingredients />} />
+                    <Route path="dishes" element={<Dishes />} />
+                    <Route path="plans">
+                      <Route index element={<Plans />} />
+                      <Route path="edit/:id" element={<EditPlan />} />
+                    </Route>
+                  </Route>
+                  <Route path="user" element={<User />}>
+                    <Route path="profile" element={<Profile />} />
+                    <Route path="change-password" element={<ChangePassword />} />
                   </Route>
                 </Route>
-                <Route path="user" element={<User />}>
-                  <Route path="profile" element={<Profile />} />
-                  <Route path="change-password" element={<ChangePassword />} />
-                </Route>
-              </Route>
-            </Routes>
+              </Routes>
+            </Suspense>
           </BrowserRouter>
           <Toast />
         </Provider>
