+++ src/planner/plans/cards.tsx
@@ -1,7 +1,7 @@
 import { Card, CardBody, CardFooter, CardHeader, Divider } from "@nextui-org/react";
 import DeleteIcon from "../../assets/deleteIcon";
 import EyeIcon from "../../assets/eyeIcon";
-import { TPlans } from "./types";
+import { TPlans } from "../../common/types";
 
 type Props = {
   data: TPlans[];
