+++ src/planner/plans/createForm.tsx
@@ -1,10 +1,12 @@
 import { Button, Checkbox, Input, ModalBody, ModalFooter, ModalHeader } from "@nextui-org/react";
 import { useFormik } from "formik";
 import * as yup from "yup";
-import { TCreatePlanBase } from "./types";
+import { TCreatePlanBase } from "../../common/types";
 
 export const schema = yup.object({
   name: yup.string().required("Name is required"),
+  isPrivate: yup.boolean(),
+  isActive: yup.boolean(),
 });
 
 type Props = {
