+++ src/planner/plans/edit/constants.ts
@@ -1,29 +1,5 @@
+import { TPlans } from "../../../common/types";
 import { defaultUser } from "../../../constants";
-import { TPlans } from "../types";
-import { EMealType } from "./types";
-
-export const days = [
-  "Monday",
-  "Tuesday",
-  "Wednesday",
-  "Thursday",
-  "Friday",
-  "Saturday",
-  "Sunday",
-] as const;
-
-export const mealTypeMap: Record<keyof typeof EMealType, EMealType> = {
-  wakeup: EMealType.wakeup,
-  breakfast: EMealType.breakfast,
-  midmorning: EMealType.midmorning,
-  lunch: EMealType.lunch,
-  midafternoon: EMealType.midafternoon,
-  dinner: EMealType.dinner,
-  bedtime: EMealType.bedtime,
-};
-
-// mealOrder = ["wakeup", "breakfast", "midmorning", ...] 
-export const mealOrder = Object.keys(EMealType) as (keyof typeof EMealType)[];
 
 export const defaultPlan: TPlans = {
   name: "",
