+++ src/planner/plans/edit/mobileView.tsx
@@ -1,8 +1,8 @@
 import { Accordion, AccordionItem, Button } from "@nextui-org/react";
 import PlusIcon from "../../../assets/plus";
-import { days } from "./constants";
-import MealCards from "./mealCards";
-import { TDays } from "./types";
+import MealCards from "../../../common/mealCards";
+import { TDays } from "../../../common/types";
+import { days } from "../../../constants";
 
 type Props = {
   openCreatePlanModal: (days: TDays) => void;
