+++ src/planner/plans/edit/types.ts
@@ -1,28 +1,5 @@
-import { days } from "./constants";
-
-export enum EMealType {
-  "wakeup" = "Wake Up",
-  "breakfast" = "Breakfast",
-  "midmorning" = "Mid Morning",
-  "lunch" = "Lunch",
-  "midafternoon" = "Mid Afternoon Snack",
-  "dinner" = "Dinner",
-  "bedtime" = "Bed Time",
-}
-
-export type MealTypeKey = keyof typeof EMealType;
-
-export type TMeal = {
-  mealType: MealTypeKey;
-  name: string;
-  dishes: TMealDishBase[];
-};
-
-export type TDays = (typeof days)[number];
-
-export type TMealDishBase = {
-  dish: { _id: string; name: string };
-};
+import { TDays, TMealDishBase } from "../../../common/types";
+import { EMealType } from "../../../constants";
 
 export type TCreateMealBase = {
   mealType: EMealType | string;
