+++ src/planner/plans/types.ts
@@ -1,23 +1,4 @@
-import { TUser } from "../../common/types";
-import { TDays, TMeal } from "./edit/types";
-
-export type TCreatePlanBase = {
-  name: string;
-  isPrivate: boolean;
-  isActive: boolean;
-};
-
-export type TPlansBase = TCreatePlanBase & {
-  meals: { [K in TDays]?: TMeal[] };
-};
-
-export type TPlans = TPlansBase & {
-  _id: string;
-  createdAt: string;
-  createdBy: TUser;
-  updatedAt: string;
-  updatedBy: TUser;
-};
+import { TPlans } from "../../common/types";
 
 export type TPlansResponse = {
   data: TPlans[];
